Improved vessel spawning.
Most of the situations giving rise to exceptions have been
eliminated. This was achieved by not using the exploder system, killing
off all the vessels, then killing off all the parts of the currently
focused vessel which somehow survives being killed in the first part.

Additionally, the autopilot issue is fixed by toggling the autopilots
twice.