Fix for allAspect and maxOffBoresight to work as intended:
allAspect = true or false //if true it simulates a modern IR missile with "uncaged lock" ability. Even if the target is not within boresight fov, it can be locked by radar and the target information transfered to the missile. It will then try to lock on with the heat seeker. If false, it is an older missile which requires a direct "in boresight" lock. Can be set to true for radar missiles to enable closer to maxOffBoresight shots (see below).

maxOffBoresight = float angle // This controls how far in degrees off the missiles centerline the radar or heat seeker can look to track the target. Also controls launch conditions. Missiles are by default allowed to launch when the angle to target is 0.35*maxOffBoresight. If allAspect = true or the target is landed/splashed, they are allowed to launch when the angle to target is 0.75*maxOffBoresight.

Add grace period to CheckMiss() to allow missiles time to turn towards the target after being launched