SteerMode improvements.
Add a "post-avoidance cool-down" slider to the pilot AI to increase pitch away from terrain for a short period after avoiding terrain.
Add a Manoeuvering steer mode similar to NormalFlight steer mode, but with Aiming mode's roll behaviour.
Switch a number of inappropriate uses of Aiming steer mode to Manoeuvering mode.
Add a green debug line for showing the pitch target direction the AI is aiming for (which is not necessarily the same as the position the AI is told to go to).